+++ packages/core/src/types/displayArea.ts
@@ -1,8 +1,13 @@
+import InterpolationType from '../enums/InterpolationType';
+
 type DisplayArea = {
+  type?: 'SCALE' | 'FIT';
+  scale?: number;
+  interpolationType?: InterpolationType;
   imageArea?: [number, number]; // areaX, areaY
   imageCanvasPoint?: {
     imagePoint: [number, number]; // imageX, imageY
-    canvasPoint: [number, number]; // canvasX, canvasY
+    canvasPoint?: [number, number]; // canvasX, canvasY
   };
   storeAsInitialCamera?: boolean;
 };