+++ packages/core/src/utilities/index.ts
@@ -80,6 +80,7 @@ import * as windowLevel from './windowLevel';
 import * as colormap from './colormap';
 import * as transferFunctionUtils from './transferFunctionUtils';
 import * as cacheUtils from './cacheUtils';
+import * as color from './color';
 
 export {
   eventListener,
@@ -164,4 +165,5 @@ export {
   getViewportImageIds,
   getRandomSampleFromArray,
   getVolumeId,
+  color,
 };