+++ packages/streaming-image-volume-loader/src/helpers/getDynamicVolumeInfo.ts
@@ -7,10 +7,11 @@ import splitImageIdsBy4DTags from './splitImageIdsBy4DTags';
  * @returns 4D series infos
  */
 function getDynamicVolumeInfo(imageIds) {
-  const timePoints = splitImageIdsBy4DTags(imageIds);
+  const { imageIdsGroups: timePoints, splittingTag } =
+    splitImageIdsBy4DTags(imageIds);
   const isDynamicVolume = timePoints.length > 1;
 
-  return { isDynamicVolume, timePoints };
+  return { isDynamicVolume, timePoints, splittingTag };
 }
 
 export default getDynamicVolumeInfo;